requirements.txt
opencv-python==4.8.1.78
mediapipe==0.10.7
numpy==1.24.3
scipy==1.10.3
pygame==2.5.2
imutils==0.5.4
pyttsx3==2.90
pandas==2.0.3
PyYAML==6.0
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0